Move Reviver out of JSONReviver.ts
